Move pkg/ice to internal/ice
Avoid exposing any non-standard API.
