+++ icecandidate_test.go
@@ -5,7 +5,7 @@ import (
 	"testing"
 
 	"github.com/pions/sdp/v2"
-	"github.com/pions/webrtc/pkg/ice"
+	"github.com/pions/webrtc/internal/ice"
 	"github.com/stretchr/testify/assert"
 )
