+++ icetransportstate.go
@@ -1,6 +1,6 @@
 package webrtc
 
-import "github.com/pions/webrtc/pkg/ice"
+import "github.com/pions/webrtc/internal/ice"
 
 // ICETransportState represents the current state of the ICE transport.
 type ICETransportState int
