+++ internal/ice/errors.go
