+++ internal/ice/networktype.go
