+++ internal/ice/networktype_test.go
