+++ internal/ice/util.go
