+++ peerconnection.go
@@ -14,7 +14,7 @@ import (
 
 	"github.com/pions/rtcp"
 	"github.com/pions/sdp/v2"
-	"github.com/pions/webrtc/pkg/ice"
+	"github.com/pions/webrtc/internal/ice"
 	"github.com/pions/webrtc/pkg/logging"
 	"github.com/pions/webrtc/pkg/rtcerr"
 	"github.com/pkg/errors"
